flask_app/requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
flask-marshmallow==0.15.0
flask-restful==0.3.9
celery==5.2.7
redis==4.5.4
python-dotenv==1.0.0
flask-redis==0.4.0
pytest-flask==1.2.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2==2.9.6
marshmallow-sqlalchemy==0.29.0
